% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsws-package.R
\docType{package}
\name{dsws-package}
\alias{dsws}
\alias{dsws-package}
\title{dsws: Bayesian Inference Workshop}
\description{
This package contains materials needed for a data science workshop organized by good job games. Check science.goodjobgames.com for futher information.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
