% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_simulated_retry}
\alias{get_simulated_retry}
\title{FUNCTION_TITLE}
\usage{
get_simulated_retry(outputfiles, min_test_user_id, min_level_id)
}
\arguments{
\item{outputfiles}{PARAM_DESCRIPTION}

\item{min_test_user_id}{PARAM_DESCRIPTION}

\item{min_level_id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[vroom]{vroom}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
}
