% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds_table_creator}
\alias{odds_table_creator}
\title{FUNCTION_TITLE}
\usage{
odds_table_creator(train_data, test_data, stan_file)
}
\arguments{
\item{train_data}{PARAM_DESCRIPTION}

\item{test_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cmdstanr]{cmdstan_model}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}
\code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
}
