% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_the_play_vector}
\alias{score_the_play_vector}
\title{Score the Betting}
\usage{
score_the_play_vector(play_vector, scores, odds_table)
}
\arguments{
\item{play_vector}{A numerical vector contains ones and zeros}

\item{scores}{A dataframe contains user id level id and retrial_count}

\item{odds_table}{A dataframe contains user_id level_id line OU and odd}
}
\description{
This function is designed to test the play vector against the given odds.
}
\details{
The function will plot the results
}
\examples{
\dontrun{
dsws::score_the_play_vector(
 model_1_play_vector, 
 scores = dsws::scores,
 odds_table = dsws::odds_table
 )
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{count}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_path}}, \code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{position_stack}}, \code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}}, \code{\link[ggplot2]{guide_legend}}, \code{\link[ggplot2]{coord_polar}}, \code{\link[ggplot2]{ggtheme}}
\code{\link[gridExtra]{arrangeGrob}}
}
