% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_simulated_retry}
\alias{get_simulated_retry}
\title{Get Simulated Retry}
\usage{
get_simulated_retry(outputfiles, min_test_user_id, min_level_id)
}
\arguments{
\item{outputfiles}{Stan output files}

\item{min_test_user_id}{Minimum user id given in the stan model}

\item{min_level_id}{Minimum level id given in the stan model}
}
\value{
return a dataframe level_id; user_id; simulated value
}
\description{
The function designed to read stan results and return the simulated retry counts
}
\details{
The function requires to be used with provided stanfiles
}
\examples{
\dontrun{
test_users_simulated_try_model_1 <- get_simulated_retry(
 fit$output_files(), 
 MIN_TEST_USER_ID,
 MIN_TEST_LEVEL_ID
 )
}
}
\seealso{
\code{\link[vroom]{vroom}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
}
