% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_the_play_vector}
\alias{score_the_play_vector}
\title{FUNCTION_TITLE}
\usage{
score_the_play_vector(play_vector, scores, odds_table)
}
\arguments{
\item{play_vector}{PARAM_DESCRIPTION}

\item{scores}{PARAM_DESCRIPTION}

\item{odds_table}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{count}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_path}}, \code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{position_stack}}, \code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}}, \code{\link[ggplot2]{guide_legend}}, \code{\link[ggplot2]{coord_polar}}, \code{\link[ggplot2]{ggtheme}}
\code{\link[gridExtra]{arrangeGrob}}
}
